uploading notebook
Using Beautiful soup to scrape the main page of the website and getting and formatting links for each vegetable guide. Also, reading and creating beautifulsoups for each vegetable guide and starting to parse data. I will need to revisit parsing for the div with a class normal since it does not appear to be consistently formatted across vegetable guide pages.